v1_TORagent update
Back to this one again.
Default dex 3 > 2.
Dex3 made buyable for concealment (instead of flipkick or ex saber training).
Default combo limit 3 > 2 (concealment still gets 4).
Ex saber training added by default.
Poison gas gun point cost increased 1 > 1.
Poison rifle reload slowed by 20%.
CR2 rifle cost 2/1/1 > 2/1/0.

Reasoning here: Dex3 was default on every spec which allowed doing big damage while getting up. That happened due to dex changes.
Mainly the gas rifle is a bit too strong for people (you can actually just not eat all 5 shots in your face so you don't die, but that seems to be hard for people). That's why it has a slightly longer reload which it was supposed to have but was cut due to the char limit, and it now costs 2 points so a 3 life build with it is harder to make (you probably will end up not taking the 1 poison dart).
Knife now has saber training which makes it slightly stronger/easier to use, but has only a 2-hit combo by default so that the sniper and poison spammer are less strong with it, while concealment remains as good.