v7_ImpAgent rework
Imp spy is now different since some of their gear went into the scouttrooper. Now it has more unique weapons and should be more fun to play while not being slightly less of a frontline class (to be fair anyone that can play will play frontline with it anyway, but at least it no longer has an E11 and it has lower shield recharge).