Han/Greedo point costs
Reworked due to open mode hero point costs. Every ability is still something to the power of 5 so there's no sense in changing it to be exactly the same, but that might be an idea for the future.
Dash 2 made free.
QT3 made buyable.
P2 made free (to make p1 builds suck a bit less if you really want to not pick a P3. Now you atleast have a P2).
Dodge 3 made cheaper.