v1_NTroop resource regen
Now regens much less but at a somewhat quicker rate. The new regen should result in better regen from moving/shooting (it drains 2 battery, but restored 4 each 10 sec). The result is that taking damage is the main loss of battery which is harder to regen (use Enoch for that), so their damage reduction goes down each time they are being shot at and it's slower to regen than health.