unfucks the vornax cannon
adds explosive damage on shots
nerfs clip size from 10 to 6
nerfs velocity from 1 to 0.6

try this out. this class is one life with the cannon, slow, needs rage to be mobile. the cannon is a big target on a guy with bad survivability. its literally a glass cannon loadout.

if you guys dont wanna try it out, i'll get back to you with a simple trispec for just the minigun loadout + level 1 a280. idea for that is adding protect, beskar 3 and/or more HP/armor. undecided at this point