baze skill timer
skilltimer mod from 2 > 1.5

still keeping dash at a long cooldown without making it excessive. this character is in a fine state for now, and making dash a thoughtful use instead of spam whenever you can is good. this keeps in line with that + current overall power level of other gunners